requirements.txt
feedparser>=6.0.10
PyYAML>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.30.0
